Clarify config settings
I accidentally renamed the original "token" config setting to "api_key",
which is wrong; this setting is for the data API, which is available on
the user settings page. "token" was more accurate, but I still found it
confusing, so I decided on "auth_token" to better clarify the
difference.